+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
   - Initial version
 
-## 0.1.0
+## 0.0.2
 
-  - Redirect responses returns false on isSuccessful()
+  - Namespace changed to Omnipay\DataCashRedirect
+
+## 0.0.3
+
+  - Remove types in GatewayParamsTrait
+  - Hardcode isSuccessful() response to false in RedirectPurchaseResponse